Fix automatic calculation of goals for CA
If there is a zero in vals then it means that there is a perfect match, which we definitely want to use as that is almost certainly the intended goal. Previously we only looked for the minimum value *after* median-filtering vals, meaning that the zero value may not be selected as the goal! Now we use this as a fallback instead (and print a warning)